Move legacy <-> Replay Client logic into standalone hook
