+++ src/ui/components/DevTools.tsx
@@ -1,6 +1,5 @@
-import { Object } from "@replayio/protocol";
-import { Pause, ThreadFront } from "protocol/thread";
-import React, { useEffect, useState, useRef, useMemo, useContext } from "react";
+import { ThreadFront } from "protocol/thread";
+import React, { useEffect, useState, useRef, useMemo } from "react";
 import { connect, ConnectedProps } from "react-redux";
 import { useAppSelector } from "ui/setup/hooks";
 import { clearTrialExpired, createSocket } from "ui/actions/session";
@@ -34,9 +33,7 @@ import { prefs } from "ui/utils/prefs";
 import { getPaneCollapse } from "devtools/client/debugger/src/selectors";
 import { getViewMode } from "ui/reducers/layout";
 import { useTrackLoadingIdleTime } from "ui/hooks/tracking";
-import { ReplayClientContext } from "shared/client/ReplayClientContext";
-import { preCacheObject } from "bvaughn-architecture-demo/src/suspense/ObjectPreviews";
-import { stringify } from "bvaughn-architecture-demo/src/utils/string";
+import useReplayClient from "ui/hooks/useReplayClient";
 
 const Viewer = React.lazy(() => import("./Viewer"));
 
@@ -117,60 +114,7 @@ function _DevTools({
     [recording]
   );
 
-  // Wire up the ReplayClient used for the new Object Inspector component.
-  const replayClient = useContext(ReplayClientContext);
-  useEffect(() => {
-    async function initSession() {
-      const sessionId = await ThreadFront.waitForSession();
-      replayClient.configure(sessionId!);
-    }
-    initSession();
-  }, [replayClient]);
-
-  // HACK
-  // Pass data (from Pause) through to the Suspense cache used by the Object Inspector.
-  // This is required because the new ReplayClient isn't being used to fetch the data.
-  useEffect(() => {
-    const set = new Set<Object>();
-
-    let currentPause: Pause | null = null;
-
-    function onPauseObjects(objects: Object[]) {
-      const pauseId = currentPause!.pauseId!;
-
-      objects.forEach((object: Object) => {
-        if (!set.has(object)) {
-          set.add(object);
-
-          // Deeply clone object data before pre-caching it.
-          // The Pause class mutates these values by adding ValueFronts to them.
-          preCacheObject(pauseId, JSON.parse(stringify(object)));
-        }
-      });
-    }
-
-    async function onThreadCurrentPause(pause: Pause | null) {
-      if (currentPause !== null) {
-        currentPause.off("objects", onPauseObjects);
-      }
-
-      currentPause = pause;
-
-      if (pause != null) {
-        pause.on("objects", onPauseObjects);
-      }
-    }
-
-    ThreadFront.on("currentPause", onThreadCurrentPause);
-
-    return () => {
-      ThreadFront.off("currentPause", onThreadCurrentPause);
-
-      if (currentPause !== null) {
-        currentPause.off("objects", onPauseObjects);
-      }
-    };
-  }, []);
+  useReplayClient();
 
   useEffect(() => {
     import("./Viewer");
@@ -203,7 +147,7 @@ function _DevTools({
     return () => {
       clearTrialExpired();
     };
-  }, [clearTrialExpired, createSocket, recordingId, replayClient]);
+  }, [clearTrialExpired, createSocket, recordingId]);
 
   useEffect(() => {
     if (uploadComplete && loadingFinished) {
